+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.4] - 2025-04-22
+
+## Fixed
+
+- Better Color Picker support
+
 ## [2.0.3] - 2025-04-21
 
 ## Added